# Properly debounce wakeups