@normanmaurer 2019-05-27
@netty-bot test this please