@carl-mastrangelo 2019-05-28
@njhill I took a crack at NioEventLoop, but it was more complicated, and I couldn't convince myself the change was correct.  I figured I would tackle this easier case first.  I can do the KQueue version, but alas I don't have a mac to test it on.  

We should really have a JCStress test for these classes.