@normanmaurer 2019-05-28
@netty-bot test this please