@normanmaurer 2019-05-28
@carl-mastrangelo I think it is fine to just do it for `epoll` first and then port it.