@franz1981 2019-05-28
Need some time to look at the code , but according to some of the things I've found on the wake-up mechanism, I'm already +1 to this :)
Just curious...we now how have 2 different benchs to measure ping pong?
If yes maybe could worth to drop one..