@carl-mastrangelo 2019-05-28
@franz1981 what other ping pong benchmark?