@franz1981 2019-05-29
@carl-mastrangelo It basically seems to measure something "similar" (not really because it doesn't involve `sync` or to create any `Future`/`Promise`): https://github.com/netty/netty/blob/cd3254df88b60476dc04b39915d3d70c200eb6f4/microbench/src/main/java/io/netty/microbench/concurrent/BurstCostExecutorsBenchmark.java

But it avoid garbage and allows burst of messages to be sent.
If you wanna give it a try with this changes you will find something interesting, I'm sure :)