@johnou 2019-05-30
@carl-mastrangelo did you get a chance to run BurstCostExecutorsBenchmark?