@carl-mastrangelo 2019-05-30
@franz1981 I did try to run BurstCostExecutorsBenchmark several times, but was stymied by maven.   The benchmark seemingly doesn't have a dependency on epoll, making the benchmark fail.  I don't know enough about maven to fix this.