@carl-mastrangelo 2019-05-30
@johnou  & @franz1981  I was able to get a run of the test.  My change is better for all cases except the last one.

```
AFTER

# Run complete. Total time: 00:08:12

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                                          (burstLength)  (executorType)  (work)    Mode      Cnt         Score    Error  Units
BurstCostExecutorsBenchmark.test1Producer                                      1  epollEventLoop       0  sample   501614      2521.188 ± 10.901  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                  1  epollEventLoop       0  sample                517.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                  1  epollEventLoop       0  sample               2464.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                  1  epollEventLoop       0  sample               2520.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                  1  epollEventLoop       0  sample               2544.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                  1  epollEventLoop       0  sample               3164.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                 1  epollEventLoop       0  sample              13462.160           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.9999                1  epollEventLoop       0  sample              18656.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p1.00                  1  epollEventLoop       0  sample            1449984.000           ns/op
BurstCostExecutorsBenchmark.test1Producer                                      1  epollEventLoop      10  sample   498098      2523.913 ±  2.983  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                  1  epollEventLoop      10  sample                313.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                  1  epollEventLoop      10  sample               2476.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                  1  epollEventLoop      10  sample               2552.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                  1  epollEventLoop      10  sample               2680.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                  1  epollEventLoop      10  sample               3048.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                 1  epollEventLoop      10  sample              13376.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.9999                1  epollEventLoop      10  sample              20326.083           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p1.00                  1  epollEventLoop      10  sample             104192.000           ns/op
BurstCostExecutorsBenchmark.test1Producer                                     10  epollEventLoop       0  sample   487331      2562.064 ±  3.320  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                 10  epollEventLoop       0  sample               1058.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                 10  epollEventLoop       0  sample               2508.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                 10  epollEventLoop       0  sample               2604.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                 10  epollEventLoop       0  sample               2640.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                 10  epollEventLoop       0  sample               3240.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                10  epollEventLoop       0  sample              13632.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.9999               10  epollEventLoop       0  sample              22169.613           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p1.00                 10  epollEventLoop       0  sample              68992.000           ns/op
BurstCostExecutorsBenchmark.test1Producer                                     10  epollEventLoop      10  sample   453820      2678.836 ±  2.892  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                 10  epollEventLoop      10  sample               1014.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                 10  epollEventLoop      10  sample               2644.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                 10  epollEventLoop      10  sample               2740.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                 10  epollEventLoop      10  sample               2788.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                 10  epollEventLoop      10  sample               3400.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                10  epollEventLoop      10  sample              13584.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.9999               10  epollEventLoop      10  sample              18880.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p1.00                 10  epollEventLoop      10  sample              47104.000           ns/op
BurstCostExecutorsBenchmark.test2Producers                                     1  epollEventLoop       0  sample  1006007      2530.363 ±  1.894  ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.00                1  epollEventLoop       0  sample                305.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.50                1  epollEventLoop       0  sample               2496.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.90                1  epollEventLoop       0  sample               2632.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.95                1  epollEventLoop       0  sample               2732.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.99                1  epollEventLoop       0  sample               3200.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.999               1  epollEventLoop       0  sample              10079.872           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.9999              1  epollEventLoop       0  sample              19065.549           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p1.00                1  epollEventLoop       0  sample             121984.000           ns/op
BurstCostExecutorsBenchmark.test2Producers                                     1  epollEventLoop      10  sample  1002643      2558.662 ±  1.852  ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.00                1  epollEventLoop      10  sample                173.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.50                1  epollEventLoop      10  sample               2560.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.90                1  epollEventLoop      10  sample               2744.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.95                1  epollEventLoop      10  sample               2804.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.99                1  epollEventLoop      10  sample               3296.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.999               1  epollEventLoop      10  sample               8880.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.9999              1  epollEventLoop      10  sample              14915.309           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p1.00                1  epollEventLoop      10  sample             151552.000           ns/op
BurstCostExecutorsBenchmark.test2Producers                                    10  epollEventLoop       0  sample   927656      2651.944 ±  2.741  ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.00               10  epollEventLoop       0  sample                588.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.50               10  epollEventLoop       0  sample               2632.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.90               10  epollEventLoop       0  sample               3308.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.95               10  epollEventLoop       0  sample               3680.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.99               10  epollEventLoop       0  sample               4640.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.999              10  epollEventLoop       0  sample              10208.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.9999             10  epollEventLoop       0  sample              15923.749           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p1.00               10  epollEventLoop       0  sample             192768.000           ns/op
BurstCostExecutorsBenchmark.test2Producers                                    10  epollEventLoop      10  sample   988291      2547.061 ± 20.093  ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.00               10  epollEventLoop      10  sample                435.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.50               10  epollEventLoop      10  sample               2448.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.90               10  epollEventLoop      10  sample               3228.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.95               10  epollEventLoop      10  sample               3532.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.99               10  epollEventLoop      10  sample               4488.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.999              10  epollEventLoop      10  sample              10848.000           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p0.9999             10  epollEventLoop      10  sample              16773.466           ns/op
BurstCostExecutorsBenchmark.test2Producers:test2Producers·p1.00               10  epollEventLoop      10  sample            4833280.000           ns/op
BurstCostExecutorsBenchmark.test3Producers                                     1  epollEventLoop       0  sample  1963843      2110.946 ±  2.717  ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.00                1  epollEventLoop       0  sample                116.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.50                1  epollEventLoop       0  sample               2532.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.90                1  epollEventLoop       0  sample               2776.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.95                1  epollEventLoop       0  sample               2864.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.99                1  epollEventLoop       0  sample               3376.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.999               1  epollEventLoop       0  sample               9362.496           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.9999              1  epollEventLoop       0  sample              17139.699           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p1.00                1  epollEventLoop       0  sample             239104.000           ns/op
BurstCostExecutorsBenchmark.test3Producers                                     1  epollEventLoop      10  sample  2048371      1889.824 ± 16.556  ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.00                1  epollEventLoop      10  sample                122.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.50                1  epollEventLoop      10  sample               2316.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.90                1  epollEventLoop      10  sample               2740.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.95                1  epollEventLoop      10  sample               2820.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.99                1  epollEventLoop      10  sample               3044.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.999               1  epollEventLoop      10  sample               7696.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.9999              1  epollEventLoop      10  sample              15573.210           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p1.00                1  epollEventLoop      10  sample           10174464.000           ns/op
BurstCostExecutorsBenchmark.test3Producers                                    10  epollEventLoop       0  sample  1749983      3536.817 ±  4.894  ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.00               10  epollEventLoop       0  sample                609.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.50               10  epollEventLoop       0  sample               3508.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.90               10  epollEventLoop       0  sample               4440.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.95               10  epollEventLoop       0  sample               4792.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.99               10  epollEventLoop       0  sample               5776.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.999              10  epollEventLoop       0  sample              10768.256           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.9999             10  epollEventLoop       0  sample              41984.205           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p1.00               10  epollEventLoop       0  sample            1081344.000           ns/op
BurstCostExecutorsBenchmark.test3Producers                                    10  epollEventLoop      10  sample  2089449      3602.025 ±  3.282  ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.00               10  epollEventLoop      10  sample                635.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.50               10  epollEventLoop      10  sample               3560.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.90               10  epollEventLoop      10  sample               4288.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.95               10  epollEventLoop      10  sample               4608.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.99               10  epollEventLoop      10  sample               5728.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.999              10  epollEventLoop      10  sample              11168.000           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p0.9999             10  epollEventLoop      10  sample              66567.040           ns/op
BurstCostExecutorsBenchmark.test3Producers:test3Producers·p1.00               10  epollEventLoop      10  sample             822272.000           ns/op



BEFORE:



Benchmark                                                          (burstLength)  (executorType)  (work)    Mode      Cnt        Score    Error  Units
BurstCostExecutorsBenchmark.test1Producer                                      1  epollEventLoop       0  sample   745106     3370.669 ±  3.309  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                  1  epollEventLoop       0  sample               384.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                  1  epollEventLoop       0  sample              3324.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                  1  epollEventLoop       0  sample              3376.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                  1  epollEventLoop       0  sample              3392.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                  1  epollEventLoop       0  sample              4104.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                 1  epollEventLoop       0  sample             16768.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.9999                1  epollEventLoop       0  sample             21056.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p1.00                  1  epollEventLoop       0  sample            505856.000           ns/op
BurstCostExecutorsBenchmark.test1Producer                                      1  epollEventLoop      10  sample   523081     3255.676 ±  3.375  ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.00                  1  epollEventLoop      10  sample               530.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.50                  1  epollEventLoop      10  sample              3276.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.90                  1  epollEventLoop      10  sample              3372.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.95                  1  epollEventLoop      10  sample              3392.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.99                  1  epollEventLoop      10  sample              4208.000           ns/op
BurstCostExecutorsBenchmark.test1Producer:test1Producer·p0.999                 1  epollEventLoop      10  sample             17184.000           ns/op
…